+++ src/Cubes/Axes.jl
@@ -18,12 +18,12 @@ end
 function YearStepRange(start::Date,stop::Date,step::Day)
     startyear=year(start)
     startday=dayofyear(start)
-    startst=ceil(Int,startday/Float64(step))
+    startst=ceil(Int,startday/Dates.value(step))
     stopyear=year(stop)
     stopday=dayofyear(stop)
-    stopst=ceil(Int,stopday/Float64(step))
-    NPY=ceil(Int,366/Float64(step))
-    YearStepRange(startyear,startst,stopyear,stopst,Int(step),NPY)
+    stopst=ceil(Int,stopday/Dates.value(step))
+    NPY=ceil(Int,366/Dates.value(step))
+    YearStepRange(startyear,startst,stopyear,stopst,Dates.value(step),NPY)
 end
 function Base.length(x::YearStepRange)
     (-x.startst+1+x.stopst+(x.stopyear-x.startyear)*x.NPY)
@@ -43,14 +43,16 @@ end
 macro defineCatAxis(axname,eltype)
   newname=esc(Symbol(string(axname,"Axis")))
   quote
-    typealias $newname CategoricalAxis{$eltype,$(QuoteNode(axname))}
+    const $newname = _CategoricalAxis{T,$(QuoteNode(axname)),R} where R<:AbstractVector{T} where T<:$(eltype)
+    $(newname)(v::AbstractVector{T}) where T = CategoricalAxis{T,$(QuoteNode(axname)),typeof(v)}(v)
   end
 end
 
-macro defineRanAxis(axname,eltype,rantype)
+macro defineRanAxis(axname,eltype)
   newname=esc(Symbol(string(axname,"Axis")))
   quote
-    typealias $newname RangeAxis{$eltype,$(QuoteNode(axname)),$rantype}
+    const $newname = _RangeAxis{T,$(QuoteNode(axname)),R} where R<:AbstractVector{T} where T<:$(eltype)
+    $(newname)(v::AbstractVector{T}) where T = RangeAxis{T,$(QuoteNode(axname)),typeof(v)}(v)
   end
 end
 
@@ -79,7 +81,7 @@ of the aliases:
 * `MSCAxis` time step inside a year (for seasonal statistics)
 
 """
-abstract CubeAxis{T,S} <: AbstractCubeMem{T,1}
+abstract type CubeAxis{T,S} <: AbstractCubeMem{T,1} end
 
 """
     CategoricalAxis{T,S}
@@ -91,18 +93,19 @@ The default constructor is:
     CategoricalAxis(axname::String,values::Vector{T})
 
 """
-immutable CategoricalAxis{T,S} <: CubeAxis{T,S}
-  values::Vector{T}
+immutable _CategoricalAxis{T,S,RT} <: CubeAxis{T,S}
+  values::RT
 end
+const CategoricalAxis = _CategoricalAxis{T,S,RT} where RT<:AbstractVector{T} where S where T
 
 
-CategoricalAxis{T}(s::Symbol,v::Vector{T})=CategoricalAxis{T,s}(v)
-CategoricalAxis(s::AbstractString,v::Vector)=CategoricalAxis(Symbol(s),v)
+CategoricalAxis{T}(s::Symbol,v::AbstractVector{T})=CategoricalAxis{T,s,typeof(v)}(v)
+CategoricalAxis(s::AbstractString,v::AbstractVector)=CategoricalAxis(Symbol(s),v)
 
 @defineCatAxis Variable String
-@defineCatAxis SpatialPoint Tuple{Float64,Float64}
+@defineCatAxis SpatialPoint Tuple{Number,Number}
 @defineCatAxis TimeScale String
-@defineCatAxis Quantile Float64
+@defineCatAxis Quantile AbstractFloat
 
 """
     RangeAxis{T,S,R}
@@ -115,20 +118,20 @@ The default constructor is:
     RangeAxis(axname::String,values::Range{T})
 
 """
-immutable RangeAxis{T,S,R} <: CubeAxis{T,S}
+immutable _RangeAxis{T,S,R} <: CubeAxis{T,S}
   values::R
 end
-RangeAxis{T}(s::Symbol,v::Range{T})=RangeAxis{T,s,typeof(v)}(v)
+const RangeAxis = _RangeAxis{T,S,R} where R<:AbstractVector{T} where S where T
+
+RangeAxis(s::Symbol,v::AbstractVector{T}) where T = RangeAxis{T,s,typeof(v)}(v)
 RangeAxis(s::AbstractString,v)=RangeAxis(Symbol(s),v)
 
 
-typealias TimeAxis RangeAxis{Date,:Time}
-TimeAxis(r)=RangeAxis(:Time,r)
 
-@defineRanAxis MSC Date YearStepRange
-@defineRanAxis Lon Float64 FloatRange{Float64}
-@defineRanAxis Lat Float64 FloatRange{Float64}
-@defineRanAxis TimeH DateTime StepRange{DateTime,Base.Dates.Minute}
+@defineRanAxis MSC TimeType
+@defineRanAxis Lon Number
+@defineRanAxis Lat Number
+@defineRanAxis Time TimeType
 
 Base.length(a::CubeAxis)=length(a.values)
 
@@ -158,7 +161,7 @@ function axVal2Index{T<:Date,S,F<:YearStepRange}(a::RangeAxis{T,S,F},v::Date)
   r = (y-a.values.startyear)*a.values.NPY + d÷a.values.step + 1
   return max(1,min(length(a.values),r))
 end
-axVal2Index{T,S,F<:FloatRange}(axis::RangeAxis{T,S,F},v;fuzzy::Bool=false)=min(max(round(Int,(v-first(axis.values))/step(axis.values))+1,1),length(axis))
+axVal2Index{T,S,F<:StepRangeLen}(axis::RangeAxis{T,S,F},v;fuzzy::Bool=false)=min(max(round(Int,(v-first(axis.values))/step(axis.values))+1,1),length(axis))
 function axVal2Index(axis::CategoricalAxis{String},v::String;fuzzy::Bool=false)
   r=findfirst(axis.values,v)
   if r==0
@@ -198,7 +201,7 @@ end
 "Fallback method"
 findAxis(a,axlist)=0
 
-getSubRange(x::CubeAxis,i)=x[i],nothing
+getSubRange(x::CubeAxis,i)=x.values[i],nothing
 getSubRange(x::TimeAxis,i)=view(x,i),nothing
 
 macro caxis_str(s)