+++ src/Proc/OnlineStats.jl
@@ -8,22 +8,22 @@ import ...CABLABTools.totuple
 export DATfitOnline
 
 
-function DATfitOnline{T<:OnlineStat{OnlineStats.ScalarInput}}(xout::AbstractArray{T},maskout,xin,maskin,cfun)
+function DATfitOnline{T<:Series{0}}(xout::AbstractArray{T},maskout,xin,maskin,cfun)
     for (mi,xi) in zip(maskin,xin)
         (mi & MISSING)==VALID && fit!(xout[1],xi)
     end
 end
 
-function DATfitOnline{T<:OnlineStat{OnlineStats.ScalarInput}}(xout::AbstractArray{T},maskout,xin,maskin,splitmask,msplitmask,cfun)
+function DATfitOnline{T<:Series{0}}(xout::AbstractArray{T},maskout,xin,maskin,splitmask,msplitmask,cfun)
   for (mi,xi,si,m2) in zip(maskin,xin,splitmask,msplitmask)
-      ((mi | m2) & MISSING)==VALID && fit!(xout[cfun(si)],xi)
+      ((mi | m2) & MISSING)==VALID && fit!(xout[cfun(si)],Float64(xi))
   end
 end
 
-function DATfitOnline{T<:OnlineStat{OnlineStats.VectorInput},U}(xout::AbstractArray{T},maskout,xin::AbstractArray{U},maskin,cfun)
+function DATfitOnline{T<:Series{1},U}(xout::AbstractArray{T},maskout,xin::AbstractArray{U},maskin,cfun)
     offs=1
     offsinc=size(xin,1)
-    xtest=zeros(U,offsinc)
+    xtest=zeros(Float64,offsinc)
     mtest=zeros(UInt8,offsinc)
     for offsin=1:offsinc:length(xin)
         for j=1:offsinc
@@ -34,7 +34,7 @@ function DATfitOnline{T<:OnlineStat{OnlineStats.VectorInput},U}(xout::AbstractAr
     end
 end
 
-function DATfitOnline{T<:OnlineStat{OnlineStats.VectorInput},U}(xout::AbstractArray{T},maskout,xin::AbstractArray{U},maskin,splitmask,msplitmask,cfun)
+function DATfitOnline{T<:Series{1},U}(xout::AbstractArray{T},maskout,xin::AbstractArray{U},maskin,splitmask,msplitmask,cfun)
   offs=1
   offsinc=size(xin,1)
   xtest=zeros(U,offsinc)
@@ -46,20 +46,20 @@ function DATfitOnline{T<:OnlineStat{OnlineStats.VectorInput},U}(xout::AbstractAr
   end
 end
 
-function finalizeOnlineCube{T<:OnlineStat,N}(c::CubeMem{T,N})
-    CubeMem(c.axes,map(i->nobs(i)>0 ? OnlineStats.value(i) : NaN,c.data),c.mask)
+function finalizeOnlineCube(c::CubeMem)
+    CubeMem(c.axes,map(i->nobs(i)>0 ? OnlineStats.value(i)[1] : NaN,c.data),c.mask)
 end
 
-function finalizeOnlineCube{T<:OnlineStats.CovMatrix,CT,S}(c::CubeMem{T},varAx::CubeAxis{CT,S})
+function finalizeOnlineCube(c::CubeMem,varAx::CubeAxis, statType::Type{T}) where {T<:CovMatrix}
   nV=length(varAx)
   cout=zeros(Float32,nV,nV,size(c.data)...)
   maskout=zeros(UInt8,nV,nV,size(c.data)...)
   cout2 = zeros(Float32,nV,size(c.data)...)
   maskout2=zeros(UInt8,nV,size(c.data)...)
   for ii in CartesianRange(size(c.data))
-    cout[:,:,ii]=OnlineStats.value(c.data[ii])
+    cout[:,:,ii]=OnlineStats.value(c.data[ii])[1]
     maskout[:,:,ii]=c.mask[ii]
-    cout2[:,ii]=OnlineStats.mean(c.data[ii])
+    cout2[:,ii]=OnlineStats.mean(c.data[ii].stats[1])
     maskout2[:,ii]=c.mask[ii]
   end
   varAx1 = isa(varAx,CategoricalAxis) ? CategoricalAxis(string(axname(varAx)," 1"),varAx.values) : RangeAxis(string(axname(varAx)," 1"),varAx.values)
@@ -67,12 +67,12 @@ function finalizeOnlineCube{T<:OnlineStats.CovMatrix,CT,S}(c::CubeMem{T},varAx::
   CubeMem(CubeAxis[varAx1,varAx2,c.axes...],cout,maskout),CubeMem(CubeAxis[varAx,c.axes...],cout2,maskout2)
 end
 
-function finalizeOnlineCube{T<:OnlineStats.KMeans,N}(c::CubeMem{T,N},varAx::CubeAxis)
-  nV,nC=size(c.data[1].value)
+function finalizeOnlineCube(c::CubeMem,varAx::CubeAxis,statType::Type{T}) where {T<:KMeans}
+  nV,nC=size(OnlineStats.value(c.data[1])[1])
   cout=zeros(Float32,nV,nC,size(c.data)...)
   maskout=zeros(UInt8,nV,nC,size(c.data)...)
   for ii in CartesianRange(size(c.data))
-    cout[:,:,ii]=OnlineStats.value(c.data[ii])
+    cout[:,:,ii]=OnlineStats.value(c.data[ii])[1]
     maskout[:,:,ii]=c.mask[ii]
   end
   classAx=CategoricalAxis("Class",["Class $i" for i=1:nC])
@@ -87,19 +87,18 @@ function getGenFun{T<:OnlineStats.KMeans}(f::Type{T},nclass,startVal,d)
   return i->begin a=f(d,nclass);a.value[:]=startVal;a end
 end
 function getGenFun{T<:OnlineStats.CovMatrix}(f::Type{T},pargs...)
-  return i->f(pargs[1],EqualWeight())
+  return i->f(pargs[1])
 end
 
 getFinalFun{T<:OnlineStat}(f::Type{T},funargs...)=finalizeOnlineCube
-getFinalFun{T<:OnlineStats.KMeans}(f::Type{T},funargs...)=c->finalizeOnlineCube(c,funargs[1])
-getFinalFun{T<:OnlineStats.CovMatrix}(f::Type{T},funargs...)=c->finalizeOnlineCube(c,funargs[1])
-
+getFinalFun{T<:OnlineStats.KMeans}(f::Type{T},funargs...)=c->finalizeOnlineCube(c,funargs[1],T)
+getFinalFun{T<:OnlineStats.CovMatrix}(f::Type{T},funargs...)=c->finalizeOnlineCube(c,funargs[1],T)
 
 function mapCube{T<:OnlineStat}(f::Type{T},cdata::AbstractCubeData,pargs...;by=CubeAxis[],max_cache=1e7,cfun=identity,outAxis=nothing,MDAxis=Void,kwargs...)
   inAxes=axes(cdata)
   #Now analyse additional by axes
   inaxtypes=map(typeof,inAxes)
-  if issubtype(T,OnlineStat{OnlineStats.VectorInput})
+  if issubtype(T,OnlineStat{1})
     MDAxis<:Void && error("$T Requires a Vector Input, you have to specify the MDAxis keyword argument.")
     issubtype(MDAxis,CubeAxis) || error("MDAxis must be an Axis type")
     mdAx = getAxis(MDAxis,inAxes)
@@ -112,7 +111,7 @@ function mapCube{T<:OnlineStat}(f::Type{T},cdata::AbstractCubeData,pargs...;by=C
     ia1 = CubeAxis[]
     funargs = ()
   end
-  function interpretBycubes(x::Union{String,DataType},c)
+  function interpretBycubes(x::Union{String,Type},c)
     i=findAxis(x,axes(c))
     axes(c)[i]
   end
@@ -162,7 +161,15 @@ function mapCube{T<:OnlineStat}(f::Type{T},cdata::AbstractCubeData,pargs...;by=C
     indims=length(bycubes)==0 ? totuple(ia1) : (totuple(ia1),totuple(CubeAxis[]))
   end
   outBroad=map(typeof,outBroad)
-  return mapCube(DATfitOnline,indata,cfun;outtype=(typeof(getGenFun(f,pargs...)(f)),),indims=indims,outdims=(outdims,),outBroadCastAxes=(outBroad,),finalizeOut=(getFinalFun(f,funargs...),),genOut=(getGenFun(f,pargs...),),kwargs...)
+  fout(x) = Series(EqualWeight(),getGenFun(f,pargs...)(x))
+  return mapCube(DATfitOnline,indata,cfun;
+    outtype=(typeof(fout(f)),),
+    indims=indims,outdims=(outdims,),
+    outBroadCastAxes=(outBroad,),
+    finalizeOut=(getFinalFun(f,funargs...),),
+    genOut=(fout,),
+    kwargs...
+)
 end
 
 include("OnlinePCA.jl")